+++ src/collapse_docs/providers/python_provider.ts
@@ -1,0 +1,6 @@
+import { DocFoldingProvider } from './doc_folding_provider';
+
+export class PythonFoldingProvider extends DocFoldingProvider {
+    languageIds = ['python'];
+    protected docRegex = /("""|''')([\s\S]*?)(\1)/g;
+}