+++ CHANGELOG.md
@@ -2,10 +2,17 @@
 
 All notable changes to this project will be documented in this file.
 
-## Unreleased
+## [1.4.1] - 2026-06-01 
 ### Added
 - Support for JSDoc folding in JSX (`.jsx`) and TSX (`.tsx`) files.
 - New **Toggle Collapse Docs** command with a default `Ctrl+Shift+Alt+D` keybinding.
+- No need to change in settings color of hided lines highlight. It's done automatically.
+- Fixed several bugs.
+- Improved performence.
+- Removed unnecessary functionalities:
+  - Collapse/uncollapse context menu buttons. <- With toggle functionlity those are not needed.
+  - Generating JsDocs <- Not really needed, thousands of extensions will do it way better.
+  - Generating function in module.export = {} <- Created only for one very specific legacy project, not needed anymore.
 
 ## [1.3.0] - 2025-05-20
 ### Added